requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.6
idna==2.8
pkg-resources==0.0.0
pycodestyle==2.5.0
PyGithub==1.43.8
PyJWT==1.7.1
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
wrapt==1.11.2
